Added new TickerCache (though it still needs work) Updated database model - using ticker_id instead of symbol in many places Fixed bugs with incorrect size data type (int instead of long - though API says int is fine) Fixed issue with data loader manager not exiting on exceptions Renamed project root to nevex-investing

+++ investing-external-data-sources/src/main/java/com/nevex/roboinvesting/api/ApiStockPrice.java
@@ -18,7 +18,7 @@ public abstract class ApiStockPrice implements Comparable<ApiStockPrice> {
 
     public abstract  BigDecimal getClose();
 
-    public abstract  Integer getVolume();
+    public abstract  Long getVolume();
 
     public abstract  BigDecimal getAdjOpen();
 
@@ -28,7 +28,7 @@ public abstract class ApiStockPrice implements Comparable<ApiStockPrice> {
 
     public abstract  BigDecimal getAdjClose();
 
-    public abstract  Integer getAdjVolume();
+    public abstract  Long getAdjVolume();
 
     public abstract  BigDecimal getDivCash();
